pyproject.toml
[tool.poetry]
name = "websocks"
version = "0.3.0"
description = "A websocket-based socks5 proxy."
readme = "README.md"
authors = ["abersheeran <me@abersheeran.com>"]
license = "MIT"

homepage = "https://github.com/abersheeran/websocks"
repository = "https://github.com/abersheeran/websocks"
documentation = "https://github.com/abersheeran/websocks/wiki"
classifiers=[
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

packages = [
    { include = "websocks" },
]

[tool.poetry.scripts]
websocks =  'websocks.commands:main'

[tool.poetry.dependencies]
python = "^3.7"
uvloop = {version = "*",markers = "sys_platform != 'win32' and sys_platform != 'cygwin' and platform_python_implementation != 'pypy'"}
websockets = "^8.1"
cryptography = "^2.8"
asyncio-socks5 = "^0.2.11"
click = "^7.0"
pyyaml = "^5.3.1"
typing-extensions = { version = "^3.7.4.1", python= "<3.8" }
aiodns = "^2.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
black = "^19.10b0"
pytest = "^5.3.2"
pysocks = "^1.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
